refactor(covid-api): cleanup components
