--- apps/covid/src/components/Card.svelte
@@ -1,20 +1,0 @@
-<script>
-  import { Box } from '@studiobear/designspek-components'
-
-  export let style = $$props.style || {}
-  export let theme = $$props.theme || { colors: { muted: '#C0D0D1' } }
-  let compStyles = {
-    mx: '0.5rem',
-    my: '0.5rem',
-    brd: '1px solid',
-    brdCol: theme.colors.muted,
-    borderRadius: '0.5rem',
-    px: '1rem',
-    py: '1rem',
-    ...style,
-  }
-</script>
-
-<Box {...$$props} style={compStyles}>
-  <slot />
-</Box>
