--- apps/covid/src/components/CardBody.svelte
@@ -1,12 +1,0 @@
-<script>
-  import { Box } from '@studiobear/designspek-components'
-
-  export let style = {}
-  let compStyles = {
-    ...style,
-  }
-</script>
-
-<Box {...$$props} style={compStyles}>
-  <slot />
-</Box>
