+++ apps/covid/src/components/index.js
@@ -1,12 +1,7 @@
 import Nav from './Nav.svelte'
 import SSR from './styleSSR.svelte'
-import OverviewBoxGlobal from './OverviewBoxGlobal.svelte'
-import OverviewBoxLocal from './OverviewBoxLocal.svelte'
 import Icons from './Icons.svelte'
 import Icon from './Icon.svelte'
-import Card from './Card.svelte'
-import CardHead from './CardHead.svelte'
-import CardBody from './CardBody.svelte'
 import Overlay from './Overlay.svelte'
 import Loading from './Loading.svelte'
 
@@ -15,13 +10,8 @@ export * from './Stats'
 export {
   Nav,
   SSR,
-  OverviewBoxGlobal,
-  OverviewBoxLocal,
   Icons,
   Icon,
-  Card,
-  CardHead,
-  CardBody,
   Overlay,
   Loading,
 }
